change dtype

+++ jax_quant_finance/black_sholes/vanilla_prices_test.py
@@ -167,13 +167,13 @@ def test_price_vol_and_expiry_scaling():
             strikes=strikes,
             expiries=expiries,
             forwards=forwards,
-            dtype=jnp.float64)
+            dtype=jnp.dtype)
     scaled_prices = option_price(
             volatilities=volatilities * scaling,
             strikes=strikes,
             expiries=expiries / scaling / scaling,
             forwards=forwards,
-            dtype=jnp.float64)
+            dtype=jnp.dtype)
     assert jnp.allclose(base_prices, scaled_prices, 1e-10)
 
 def test_option_prices_detailed_discount():
